Simplify SM
